refactor: to mutualize element from progress indicators
