+++ OUDS/Core/Components/Sources/Indicators/ProgressIndicator/Internal/CircularProgressIndicator/CircularProgressIndicatorDeterminateView.swift
@@ -94,11 +94,11 @@ struct CircularProgressIndicatorDeterminateView: View {
         #if os(watchOS) || os(visionOS)
             .onChange(of: progress) { _, newValue in
                 apply(newValue: newValue)
-            }
+        }
         #else
-            .onChange(of: progress) { newValue in
-                apply(newValue: newValue)
-            }
+        .onChange(of: progress) { newValue in
+            apply(newValue: newValue)
+        }
         #endif
     }
