+++ OUDS/Core/Components/Sources/Indicators/ProgressIndicator/Internal/LinearProgressIndicator/LinearProgressIndicatorView.swift
@@ -37,7 +37,7 @@ struct LinearProgressIndicatorView: View {
 
             helperText
         }
-        .modifier(LinearProgressAccessibilityModifier(configuration: configuration))
+        .modifier(ProgressIndicatorAccessibilityModifier(progress: configuration.progress, accessibilityLabel: configuration.accessibilityLabel))
     }
 
     // MARK: - Sub-views
@@ -121,77 +121,5 @@ struct LinearProgressIndicatorView: View {
             ? theme.progressIndicator.borderRadiusRounded
             : theme.progressIndicator.borderRadiusDefault
         return (effectiveRadius > 0) ? .round : .butt
-    }
-}
-
-// MARK: - Linear Progress Accessibility Modifier
-
-/// Applies accessibility traits and values on the linear progress indicator.
-///
-/// - **Determinate**: exposes the percentage value (e.g. *"75 percent"*) as `accessibilityValue`, with
-///   the optional `helperText` as `accessibilityLabel`. The `.updatesFrequently` and `.isStaticText`
-///   traits are set so assistive technologies read the changing progress. On iOS 17+ /
-///   macOS 14+ / visionOS 1+ / watchOS 10+ / tvOS 17+, the element opts out of user interaction via
-///   `accessibilityRespondsToUserInteraction(false)` so that Full Keyboard Access does not focus it.
-/// - **Indeterminate without helper text**: hidden from VoiceOver (`.accessibilityHidden(true)`),
-///   which also excludes it from Full Keyboard Access navigation.
-/// - **Indeterminate with helper text**: the helper text is exposed as `accessibilityLabel` so the
-///   user still understands the context, but no value is exposed (no measurable progress).
-private struct LinearProgressAccessibilityModifier: ViewModifier {
-
-    let configuration: LinearProgressIndicatorConfiguration
-
-    func body(content: Content) -> some View {
-        if let progress = configuration.progress {
-            let percent = Int((progress * 100).rounded())
-
-            let determinate = content
-                .accessibilityElement(children: .ignore)
-                .accessibilityAddTraits([.updatesFrequently, .isStaticText])
-                .accessibilityLabel(a11yLabel ?? "")
-                .accessibilityValue("\(percent)%")
-
-            if #available(iOS 17, macOS 14, visionOS 1, watchOS 10, tvOS 17, *) {
-                determinate.accessibilityRespondsToUserInteraction(false)
-            } else {
-                determinate
-            }
-        } else if let a11yLabel, !a11yLabel.isEmpty {
-            // Indeterminate with helper text: expose the label but no value.
-            let indeterminate = content
-                .accessibilityElement(children: .ignore)
-                .accessibilityAddTraits(.isStaticText)
-                .accessibilityLabel(a11yLabel)
-
-            if #available(iOS 17, macOS 14, visionOS 1, watchOS 10, tvOS 17, *) {
-                indeterminate.accessibilityRespondsToUserInteraction(false)
-            } else {
-                indeterminate
-            }
-        } else {
-            // Indeterminate without helper text: hidden from VoiceOver (and therefore from FKA too).
-            content.accessibilityHidden(true)
-        }
-    }
-
-    // MARK: - Accessibility Helper
-
-    private var a11yLabel: String? {
-        switch configuration {
-        case let .determinate(configuration):
-            switch configuration.helperText {
-            case let .description(description, _):
-                description
-
-            case let .percent(_, description, alignment):
-                description
-
-            case .none:
-                nil
-            }
-
-        case let .indeterminate(configuration):
-            configuration.helperText
-        }
     }
 }
