+++ OUDS/Core/Components/Sources/Indicators/ProgressIndicator/OUDSCircularProgressIndicator.swift
@@ -67,6 +67,19 @@ import SwiftUI
 ///
 /// The indeterminate mode is not affected by this flag: its Android Material 3 animation is intrinsic to the mode.
 ///
+/// ## Helper text
+///
+/// An optional text displayed below the circular indicator to provide context or additional information.
+/// The helper text is always centered under the indicator.
+///
+/// In **determinate** mode, use ``OUDSCircularProgressIndicator.HelperTextType`` which offers two variants:
+///
+/// - **`.description(_:)`**: Displays a simple description text without the progress percentage.
+///
+/// - **`.percent(description:spaceBefore:)`**: Displays the progress percentage with an optional description.
+///
+/// In **indeterminate** mode, the `helperText` parameter is a simple `String?`.
+///
 /// ## Code samples
 ///
 /// ```swift
@@ -82,9 +95,19 @@ import SwiftUI
 ///     // Determinate displayed instantly at its target value, without any animation
 ///     OUDSCircularProgressIndicator(progress: 0.75, animated: false)
 ///
+///     // Determinate with helper text
+///     OUDSCircularProgressIndicator(progress: 0.75, helperText: .description("Uploading..."))
+///
+///     // Determinate with percentage and description
+///     OUDSCircularProgressIndicator(progress: 0.75,
+///                                   helperText: .percent(description: "of 100 MB", spaceBefore: true))
+///
 ///     // Indeterminate
 ///     OUDSCircularProgressIndicator()
 ///     OUDSCircularProgressIndicator(status: .info)
+///
+///     // Indeterminate with helper text
+///     OUDSCircularProgressIndicator(status: .info, helperText: "Processing...")
 /// ```
 ///
 /// ## Accessibility considerations
@@ -141,6 +164,26 @@ public struct OUDSCircularProgressIndicator: View { // TODO: #409 - Update docum
     /// Embeds all configuration details for the circular progress indicator
     private let configuration: CircularProgressIndicatorConfiguration
 
+    // MARK: - Determinate Progress Indicator Helper Text
+
+    /// The helper text can be added in **determinate** indicator
+    ///
+    /// - Since: 3.0.0
+    @frozen public enum HelperTextType: Equatable {
+
+        /// The helper text with a description without any information of progress.
+        ///
+        /// - Parameter description: Text displayed
+        case description(_ description: String)
+
+        /// Displays the progress information (percentage value and `%` character with optional space before) in the helper text.
+        ///
+        /// - Parameters:
+        ///   - description: Optional description text displayed alongside the percentage.
+        ///   - spaceBefore: When `true`, adds a non-breaking space before the `%` character (e.g., "75 %"). When `false`, no space (e.g., "75%"). Defaults to `false`.
+        case percent(_ description: String? = nil, spaceBefore: Bool = false)
+    }
+
     // MARK: - Initializers
 
     /// Creates a **determinate** circular progress indicator.
@@ -155,19 +198,22 @@ public struct OUDSCircularProgressIndicator: View { // TODO: #409 - Update docum
     ///      instantly at its target value with no animation. Animations are always disabled when
     ///      `accessibilityReduceMotion` is on or when Low Power Mode is enabled, regardless of this flag.
     ///    - size: The size of the component could be adjusted if used internally by components.
+    ///    - helperText: Optional helper text displayed below the indicator. Defaults to `nil`.
     public init(progress: Double,
                 status: OUDSProgressIndicatorStatus = .neutral,
                 track: Bool = true,
                 gapSize: OUDSProgressIndicatorGapSize = .default,
                 animated: Bool = true,
-                size: CGFloat = Self.defaultSize)
+                size: CGFloat = Self.defaultSize,
+                helperText: Self.HelperTextType? = nil)
     {
         configuration = CircularProgressIndicatorConfiguration(progress: progress,
                                                                status: status,
                                                                track: track,
                                                                gapSize: gapSize,
                                                                size: size,
-                                                               animated: animated)
+                                                               animated: animated,
+                                                               helperText: helperText)
     }
 
     /// Creates an **indeterminate** circular progress indicator.
@@ -177,17 +223,20 @@ public struct OUDSCircularProgressIndicator: View { // TODO: #409 - Update docum
     ///    - track: Whether the track is displayed. Defaults to `true`.
     ///    - gapSize: The size of the gap between the indicator and the track. Defaults to ``OUDSProgressIndicatorGapSize/default``.
     ///    - size: The size of the component could be adjusted if used internally by components.
+    ///    - helperText: Optional helper text displayed below the indicator. Defaults to `nil`.
     public init(status: OUDSProgressIndicatorStatus = .neutral,
                 track: Bool = true,
                 gapSize: OUDSProgressIndicatorGapSize = .default,
-                size: CGFloat = Self.defaultSize)
+                size: CGFloat = Self.defaultSize,
+                helperText: String? = nil)
     {
         configuration = CircularProgressIndicatorConfiguration(progress: nil,
                                                                status: status,
                                                                track: track,
                                                                gapSize: gapSize,
                                                                size: size,
-                                                               animated: true)
+                                                               animated: true,
+                                                               helperText: .description(helperText ?? ""))
     }
 
     // MARK: - Body
