+++ OUDS/Core/Components/Sources/Indicators/ProgressIndicator/OUDSLinearProgressIndicator.swift
@@ -76,7 +76,7 @@ import SwiftUI
 ///
 /// An optional text displayed below the progress bar to provide context or additional information.
 ///
-/// In **determinate** mode, use ``OUDSDeterminateProgressIndicatorHelperText`` which offers two variants:
+/// In **determinate** mode, use ``OUDSLinearProgressIndicator.HelperTextType`` which offers two variants:
 ///
 /// - **`.description(_:)`**: Displays a simple description text without the progress percentage.
 ///   The `alignment` parameter controls horizontal alignment (`.center`, `.start`, or `.end`).
@@ -228,6 +228,30 @@ public struct OUDSLinearProgressIndicator: View { // TODO: #1509 - Add hyperlink
     /// Embeds all configuration details for the linear progress indicator.
     private let configuration: LinearProgressIndicatorConfiguration
 
+    // MARK: - Determinate Progress Indicator Helper Text
+
+    /// The helper text can be added in **determinate** indicator
+    ///
+    /// - Since: 3.0.0
+    @frozen public enum HelperTextType: Equatable {
+
+        /// The helper text with a description without any information of progress.
+        ///
+        /// - Parameters:
+        ///   - description: Text displayed
+        ///   - alignment: Controls the horizontal alignment of the description text. Defaults to `.center`.
+        case description(_ description: String, alignment: OUDSLinearProgressIndicatorHelperTextAlignment = .center)
+
+        /// Displays the progress information (percentage value and `%` character with optional space before)
+        /// in the helper text according to the alignment.
+        ///
+        /// - Parameters:
+        ///   - spaceBefore: When `true`, adds a non-breaking space before the `%` character (e.g., "75 %"). When `false`, no space (e.g., "75%"). Defaults to `false`.
+        ///   - description: Optional description text displayed alongside the percentage. Ignored when `alignment` is `.center`.
+        ///   - alignment: The horizontal alignment of the helper text. Defaults to `.center`.
+        case percent(spaceBefore: Bool = false, description: String? = nil, alignment: OUDSLinearProgressIndicatorHelperTextAlignment = .center)
+    }
+
     // MARK: - Initializers
 
     /// Creates a **determinate** linear progress indicator.
@@ -249,7 +273,7 @@ public struct OUDSLinearProgressIndicator: View { // TODO: #1509 - Add hyperlink
                 status: OUDSProgressIndicatorStatus = .neutral,
                 track: Bool = true,
                 stopIndicator: Bool = false,
-                helperText: OUDSDeterminateProgressIndicatorHelperText? = .percent(spaceBefore: true, alignment: .center),
+                helperText: Self.HelperTextType? = .percent(spaceBefore: true, alignment: .center),
                 gapSize: OUDSProgressIndicatorGapSize = .default,
                 animated: Bool = true)
     {
